data/pyproject.toml
[project]
name = "ygo-image-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "requests>=2.32.5",
]
